Finally Fix the Fixups
**************** That was pain (2d)
Thanks AVR for having the same word-addressed PC.
All references to PC in memory and instrucions are now shifted by >>2.
Values in data memory (indirect) are distinguished from normal values and different fixup is applied.
One fixup additionaly writes to next byte (for instruction field), and other is for memory and just shifts the value.
Doesn't work in assembly parser, but I dont care for now.