Finnaly fix the ASM Relocations
The master workaround.
Additional VK only for Constants that are PC_REF is added, and @pcref
variant in artificially added in instruction printing. This way it only
affects the .<type> definitions and not internal label, which can't
parse variants.
Fixed parsing variants from assembly, they are mapped to correct
relocation from generic fixup.