Add branch optimization hooks
analyze, insert, remove and invert condition of branches.
Found because of accidetnal assert failure on missing insert,
and turns out, that it is exactly what I needed (especially analyzeBranch).
It removes all of those jump to next line things ^^. Very nice that opts to
branches are working good. Based on Lanai.